LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/table_grow64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds (table_grow64.wast:1) assert_equals: expected false but got true
PASS #3 Reinitialize the default imports
FAIL #4 Test that WebAssembly compilation succeeds (table_grow64.wast:1) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't validate: table.get index to type I64 expected I32, in function at index 0 at {loc} expected true got false
FAIL #5 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #6 Test that a WebAssembly code returns a specific result (table_grow64.wast:12) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:10:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #7 Test that a WebAssembly code traps (table_grow64.wast:13) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:13:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #8 Test that a WebAssembly code traps (table_grow64.wast:14) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:16:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #9 Test that a WebAssembly code returns a specific result (table_grow64.wast:16) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:19:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #10 Test that a WebAssembly code returns a specific result (table_grow64.wast:17) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:22:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #11 Test that a WebAssembly code returns a specific result (table_grow64.wast:18) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:25:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #12 Test that a WebAssembly code returns a specific result (table_grow64.wast:19) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #13 Test that a WebAssembly code returns a specific result (table_grow64.wast:20) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #14 Test that a WebAssembly code traps (table_grow64.wast:21) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:34:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #15 Test that a WebAssembly code traps (table_grow64.wast:22) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:37:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #16 Test that a WebAssembly code returns a specific result (table_grow64.wast:24) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:40:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #17 Test that a WebAssembly code returns a specific result (table_grow64.wast:25) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:43:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #18 Test that a WebAssembly code returns a specific result (table_grow64.wast:26) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:46:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #19 Test that a WebAssembly code returns a specific result (table_grow64.wast:27) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:49:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #20 Test that a WebAssembly code returns a specific result (table_grow64.wast:28) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:52:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #21 Test that a WebAssembly code returns a specific result (table_grow64.wast:29) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:55:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #22 Test that a WebAssembly code returns a specific result (table_grow64.wast:30) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:58:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #23 Test that a WebAssembly code returns a specific result (table_grow64.wast:31) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #24 Test that a WebAssembly code returns a specific result (table_grow64.wast:32) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:64:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #25 Test that a WebAssembly code traps (table_grow64.wast:33) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:67:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
FAIL #26 Test that a WebAssembly code traps (table_grow64.wast:34) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
table_grow64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:70:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/table_grow64.wast.js:72:3 expected true got false
